# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OWC Website API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /legislation:
    get:
      operationId: listLegislation
      tags: [legislation]
      summary: List legislation
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: type
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: List of legislation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegislationListResponse"
    post:
      operationId: createLegislation
      tags: [legislation]
      summary: Create legislation entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegislationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegislationItem"
  /legislation/{id}:
    get:
      operationId: getLegislationById
      tags: [legislation]
      summary: Get a legislation item
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Legislation item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegislationItem"
        "404":
          description: Not found
    patch:
      operationId: updateLegislation
      tags: [legislation]
      summary: Update legislation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegislationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegislationItem"
    delete:
      operationId: deleteLegislation
      tags: [legislation]
      summary: Delete legislation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── TENDERS ─────────────────────────────────────────────────────────────────
...
  /faqs:
    get:
      operationId: listFaqs
      tags: [faqs]
      summary: List FAQs
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: List of FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Faq"
    post:
      operationId: createFaq
      tags: [faqs]
      summary: Create FAQ
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
  /faqs/{id}:
    get:
      operationId: getFaqById
      tags: [faqs]
      summary: Get a FAQ
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: FAQ
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
        "404":
          description: Not found
    patch:
      operationId: updateFaq
      tags: [faqs]
      summary: Update FAQ
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaqUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Faq"
    delete:
      operationId: deleteFaq
      tags: [faqs]
      summary: Delete FAQ
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── FORMS & DOWNLOADS ────────────────────────────────────────────────────────
  /forms:
    get:
      operationId: listForms
      tags: [forms]
      summary: List forms and downloads
      parameters:
        - in: query
          name: category
          schema: { type: string }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: List of forms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Form"
    post:
      operationId: createForm
      tags: [forms]
      summary: Create form/download entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
  /forms/{id}:
    get:
      operationId: getFormById
      tags: [forms]
      summary: Get a form
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Form
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
        "404":
          description: Not found
    patch:
      operationId: updateForm
      tags: [forms]
      summary: Update form
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FormUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Form"
    delete:
      operationId: deleteForm
      tags: [forms]
      summary: Delete form
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── SERVICES ─────────────────────────────────────────────────────────────────
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
  /contact/submissions:
    get:
      operationId: listContactSubmissions
      tags: [contact]
      summary: List contact submissions (staff only)
      parameters:
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: status
          schema: { type: string }
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmissionListResponse"
  /contact/submissions/{id}:
    get:
      operationId: getContactSubmission
      tags: [contact]
      summary: Get contact submission
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Submission
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"
    patch:
      operationId: updateContactSubmission
      tags: [contact]
      summary: Update contact submission status
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSubmissionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmission"

  # ─── SEARCH ──────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LegislationItem:
      type: object
      required: [id, title, type, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
        type: { type: string }
        year: { type: ["integer", "null"] }
        fileUrl: { type: ["string", "null"] }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    LegislationListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/LegislationItem"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    LegislationInput:
      type: object
      required: [title, type]
      properties:
        title: { type: string, minLength: 1 }
        slug: { type: string }
        description: { type: string }
        type: { type: string }
        year: { type: integer }
        fileUrl: { type: string }
        status: { type: string }
    LegislationUpdate:
      type: object
      properties:
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        type: { type: string }
        year: { type: integer }
        fileUrl: { type: string }
        status: { type: string }

    # ─── TENDERS ───────────────────────────────────────────────────────────────
...
    Faq:
      type: object
      required: [id, question, answer, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        question: { type: string }
        answer: { type: string }
        category: { type: ["string", "null"] }
        order: { type: integer }
        status: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    FaqInput:
      type: object
      required: [question, answer]
      properties:
        question: { type: string, minLength: 1 }
        answer: { type: string, minLength: 1 }
        category: { type: string }
        order: { type: integer }
        status: { type: string }
    FaqUpdate:
      type: object
      properties:
        question: { type: string }
        answer: { type: string }
        category: { type: string }
        order: { type: integer }
        status: { type: string }

    # ─── FORMS ─────────────────────────────────────────────────────────────────
    Form:
      type: object
      required: [id, title, status, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        fileUrl: { type: ["string", "null"] }
        fileSize: { type: ["string", "null"] }
        fileType: { type: ["string", "null"] }
        version: { type: ["string", "null"] }
        status: { type: string }
        downloadCount: { type: integer }
        createdAt: { type: string }
        updatedAt: { type: string }
    FormInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        description: { type: string }
        category: { type: string }
        fileUrl: { type: string }
        fileSize: { type: string }
        fileType: { type: string }
        version: { type: string }
        status: { type: string }
    FormUpdate:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        category: { type: string }
        fileUrl: { type: string }
        fileSize: { type: string }
        fileType: { type: string }
        version: { type: string }
        status: { type: string }

    # ─── SERVICES ──────────────────────────────────────────────────────────────
...
    ContactSubmission:
      type: object
      required: [id, name, email, subject, message, status, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        phone: { type: ["string", "null"] }
        subject: { type: string }
        message: { type: string }
        status: { type: string }
        respondedAt: { type: ["string", "null"] }
        createdAt: { type: string }
    ContactSubmissionListResponse:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContactSubmission"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
    ContactInput:
      type: object
      required: [name, email, subject, message]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string }
        phone: { type: string }
        subject: { type: string, minLength: 1 }
        message: { type: string, minLength: 1 }
    ContactSubmissionUpdate:
      type: object
      properties:
        status: { type: string }

    # ─── SEARCH ────────────────────────────────────────────────────────────────
...