# lib/api-zod/src/generated/types/faqInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OWC Website API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FaqInput {
  /** @minLength 1 */
  question: string;
  /** @minLength 1 */
  answer: string;
  category?: string;
  order?: number;
  status?: string;
}