# lib/api-zod/src/generated/types/form.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OWC Website API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Form {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  fileUrl?: string | null;
  /** @nullable */
  fileSize?: string | null;
  /** @nullable */
  fileType?: string | null;
  /** @nullable */
  version?: string | null;
  status: string;
  downloadCount?: number;
  createdAt: string;
  updatedAt: string;
}