# lib/api-zod/src/generated/types/legislationUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * OWC Website API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LegislationUpdate {
  title?: string;
  slug?: string;
  description?: string;
  type?: string;
  year?: number;
  fileUrl?: string;
  status?: string;
}